% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_by_area.R
\docType{data}
\name{annual_by_area}
\alias{annual_by_area}
\title{Population by area example data}
\format{
A data frame with 25 rows and 5 variables:
\describe{
  \item{year}{A vector of years.}
  \item{Narnia}{A vector of random numbers.}
  \item{Westeros}{A vector of random numbers.}
  \item{Arrakis}{A vector of random numbers.}
}
}
\source{
Simulated data
}
\usage{
annual_by_area
}
\description{
A dataset containing simulated metric data for package examples. This example is for four population metrics corresponding to different areas.
}
\keyword{datasets}
