% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cleaned_data.R
\name{convert_cleaned_data}
\alias{convert_cleaned_data}
\title{Convert cleaned indicator dataset to indicator data format (with metadata rows)}
\usage{
convert_cleaned_data(data, indicator_names, unit_names, extent_names)
}
\arguments{
\item{data}{A data frame or matrix containing numeric data only.}

\item{indicator_names}{A character vector of names for the top most header row (the name of the indicator). The first name corresponding to year should be blank.}

\item{unit_names}{A character vector of names for the second header row (e.g. units of measurement).}

\item{extent_names}{A character vector of names for the third header row (e.g. area or species names).}
}
\value{
A data frame containing the three metadata rows followed by the data rows.
}
\description{
This function takes a numeric data frame with year as column 1 and metric values in the remaining columns and combines it with three vectors representing different levels of header metadata (indicator, unit, and extent/sub-indicator). It returns a single data frame that is in the correct format for ingestion into the IEAnalyzeR data_prep function.
}
\examples{
# 1. Define dummy data
dat <- data.frame(
  year = 2000:2004,
  bio_spec1 = c(5, 10, 3, 7, 11),
  bio_spec2 = c(7, 3, 8, 8, 12)
)

# 2. Define header components for the data rows (ignore year)
indicator_names <- c("Biomass", "Biomass")
unit_names <- c("Count", "Count")
extent_names <- c("species A", "species B")

# 3. Call the function
final_table <- convert_cleaned_data(dat, indicator_names, unit_names, extent_names)
print(final_table)
}
