% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_data_formatted.R
\docType{data}
\name{multi_data_formatted}
\alias{multi_data_formatted}
\title{Multi metric formatted example data}
\format{
A data frame with 132 rows and 5 variables:
\describe{
  \item{date}{A vector of months.}
  \item{temp_a}{An vector of random numbers.}
  \item{salinity_a}{A vector of random numbers.}
  \item{temp_b}{An vector of random numbers.}
  \item{salinity_b}{A vector of random numbers.}
}
}
\source{
Simulated data
}
\usage{
multi_data_formatted
}
\description{
A list object containing simulated metric data for package examples. This example is for a multi metric indicator (a vector of random numbers). This has been prepped using the data_prep function, for use with IEAnalyzeR's plotting function.
}
\keyword{datasets}
