% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_unit_monthly.R
\docType{data}
\name{multi_unit_monthly}
\alias{multi_unit_monthly}
\title{Multi-unit monthly example data}
\format{
A data frame with 132 rows and 5 variables:
\describe{
  \item{date}{A vector of months.}
  \item{temp_a}{An vector of random numbers.}
  \item{salinity_a}{A vector of random numbers.}
  \item{temp_b}{An vector of random numbers.}
  \item{salinity_b}{A vector of random numbers.}
}
}
\source{
Simulated data
}
\usage{
multi_unit_monthly
}
\description{
A dataset containing simulated metric data for package examples. This example is for two metrics with different units and data in monthly format.
}
\keyword{datasets}
