% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn_obj.R
\name{plot_fn_obj}
\alias{plot_fn_obj}
\title{Title}
\usage{
plot_fn_obj(
  df_obj,
  interactive = FALSE,
  sep_ylabs = F,
  ylab_sublabel = NULL,
  facet_scales = "free",
  ncol = NULL,
  manual_ylab = NULL,
  manual_xlab = NULL,
  manual_title = NULL,
  lwd = 0.75,
  pts = F,
  pt_size = 0.75,
  fig.width = 6,
  facet_grid = F,
  xbreaks_by = NULL,
  trend = F,
  ci = T
)
}
\arguments{
\item{df_obj}{Data object produced by the "data_prep" function.}

\item{interactive}{Run plot through plotly to create an interactive version of the plot.}

\item{sep_ylabs}{Uses facet_wrap to allow the subplots to have separate y-axis labels. Only works for indicators that are one column and are multi.}

\item{ylab_sublabel}{This argument creates a sublabel for the y-axis by choosing two of the three label types. Label types include “indicator”, “unit”, and “extent”.}

\item{facet_scales}{Derived from facet_wrap scales argument : Should scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")? This controls the share vs “free” axis per panel in a faceted plot.}

\item{ncol}{How many columns should be in the faceted figure?}

\item{manual_ylab}{Manual y axis title. List multiple ylabs in order of appearance for figure with subindicators.}

\item{manual_xlab}{Manual x axis title.}

\item{manual_title}{Manual overall plot/plots title.}

\item{lwd}{Adjust the thickness of the timeseries line.}

\item{pts}{T/F Would you like points where the data are on top of the timeseries line.}

\item{pt_size}{What size would you like the points on the graph if pts=T?}

\item{fig.width}{The intended width of saved figure. Is necessary for proper spacing if using trend symbol on final plot whether saving through the function or separately.}

\item{facet_grid}{Used if you need a subset of plots organized by two categories of labels. T defaults to unit by extent, or can specify two out of the three ("indicator","unit","extent")}

\item{xbreaks_by}{Specify the interval that the x axis breaks occur.}

\item{trend}{Do you want to include trend symbols (mean & slope) on the side of the plots.}

\item{ci}{T/F If you would like cocnfidence intervals to be displayed. CI data must have been included in the data_prep stage.}
}
\value{
A plot in the indicatorTimeSeries format.
}
\description{
This function plots an indicator time series figure from data that is formated from the "data_prep" function in IEAnalyzeR.
}
