% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_data_formatted.R
\docType{data}
\name{single_data_formatted}
\alias{single_data_formatted}
\title{Single metric formatted example data}
\format{
A list obect of several dataframes:
\describe{
  \item{year}{A vector of years.}
  \item{red_fish}{A vector of random numbers.}
}
}
\source{
Simulated data
}
\usage{
single_data_formatted
}
\description{
A list object containing simulated metric data for package examples. This example is for a single metric (biomass of red fish). This has been prepped using the data_prep function, for use with IEAnalyzeR's plotting function.
}
\keyword{datasets}
