% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_metric.R
\docType{data}
\name{single_metric}
\alias{single_metric}
\title{Single metric example data}
\format{
A data frame with 25 rows and 2 variables:
\describe{
  \item{year}{A vector of years.}
  \item{red_fish}{A vector of random numbers.}
}
}
\source{
Simulated data
}
\usage{
single_metric
}
\description{
A dataset containing simulated metric data for package examples. This example is for a single metric (biomass of red fish).
}
\keyword{datasets}
