% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twocat_data_formatted.R
\docType{data}
\name{twocat_data_formatted}
\alias{twocat_data_formatted}
\title{Count and revenue data per region}
\format{
A multi-indicator data frame with data in two main categories:
\describe{
  \item{year}{A vector of years.}
  \item{rev_narnia}{A vector of random numbers.}
  \item{count_narnia}{A vector of random numbers.}
  \item{rev_wakanda}{A vector of random numbers.}
  \item{count_wakanda}{A vector of random numbers.}
}
}
\source{
Simulated data
}
\usage{
twocat_data_formatted
}
\description{
A dataset containing simulated metric data for package examples. This example is for four population metrics corresponding to different areas.
}
\keyword{datasets}
