% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R, R/data_prep_dev.R
\name{data_prep}
\alias{data_prep}
\title{Title}
\usage{
data_prep(df, trends = T, subind = FALSE, window = 5)

data_prep(df, trends = T, subind = FALSE, window = 5)
}
\arguments{
\item{df}{Dataset with top 3 rows inlcuding metadata of indicator name, unit, and subcategory.}

\item{trends}{T/F if you would like the function to calculate trends on this dataset.}

\item{subind}{How is the data categorized for sub indicators. Options "extent" or "unit".}

\item{window}{How big is the window for the trend analysis. Default is 5 years.}
}
\value{
An object with 5 datasets used in "plot_fn_obj".

An object with 5 datasets used in "plot_fn_obj".
}
\description{
This function re-formats data from the "plotIndicatorTimeSeries" csv file format, into a data object that works with plotting functions in the IEAnalyzeR function.

This function re-formats data from the "plotIndicatorTimeSeries" csv file format, into a data object that works with plotting functions in the IEAnalyzeR function.
}
