% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn_obj.R
\name{plot_fn_obj}
\alias{plot_fn_obj}
\title{Title}
\usage{
plot_fn_obj(df_obj, interactive = FALSE)
}
\arguments{
\item{df_obj}{Data object produced by the "data_prep" function.}

\item{interactive}{Run plot through plotly to create an interactive version of the plot.}
}
\value{
A plot in the indicatorTimeSeries format.
}
\description{
This function plots an indicator time series figure from data that is formated from the "data_prep" function in IEAnalyzeR.
}
